Splitted padigitale job and metrics job
