+++ metrics-report-generator/src/main/scala/it/pagopa/interop/metricsreportgenerator/util/Error.scala
@@ -4,6 +4,4 @@ import java.util.UUID
 
 object Error {
   final case class TenantNotFound(tenantId: UUID) extends Exception(s"Tenant ${tenantId.toString} not found")
-  final case class MissingActivationTimestamp(descriptorId: UUID)
-      extends Exception(s"Descriptor ${descriptorId.toString} without activatedAt field")
 }
