+++ metrics-report-generator/src/main/scala/it/pagopa/interop/metricsreportgenerator/util/FileWriters.scala
@@ -1,8 +1,7 @@
 package it.pagopa.interop.metricsreportgenerator.util
 
-import io.circe.syntax._
 import it.pagopa.interop.commons.utils.service.OffsetDateTimeSupplier
-import it.pagopa.interop.metricsreportgenerator.report.{AgreementRecord, Metric}
+import it.pagopa.interop.metricsreportgenerator.report.AgreementRecord
 import spray.json.enrichAny
 
 import java.time.format.DateTimeFormatter
@@ -12,15 +11,6 @@ import scala.concurrent.Future
 final class FileWriters(fileUtils: FileUtils, dateTimeSupplier: OffsetDateTimeSupplier) {
 
   val dtf: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss")
-
-  def paDigitaleWriter: Seq[Metric] => Future[String] = metrics => {
-    val fileName: String = s"${dateTimeSupplier.get().format(dtf)}_${UUID.randomUUID()}.ndjson"
-    fileUtils.store(
-      ApplicationConfiguration.paDigitaleContainer,
-      ApplicationConfiguration.paDigitaleStoragePath,
-      fileName
-    )(metrics.map(_.asJson.noSpaces))
-  }
 
   def agreementsJsonWriter: Seq[AgreementRecord] => Future[String] = agreementRecords => {
     val fileName: String = s"${dateTimeSupplier.get().format(dtf)}_${UUID.randomUUID()}.ndjson"
