+++ metrics-report-generator/src/main/scala/it/pagopa/interop/metricsreportgenerator/util/ReadModelQueries.scala
@@ -1,7 +1,5 @@
 package it.pagopa.interop.metricsreportgenerator.util
 
-import it.pagopa.interop.catalogmanagement.model.CatalogItem
-import it.pagopa.interop.catalogmanagement.model.persistence.JsonFormats._
 import it.pagopa.interop.commons.cqrs.service.ReadModelService
 import it.pagopa.interop.commons.utils.TypeConversions.OptionOps
 import it.pagopa.interop.metricsreportgenerator.models.{Agreement, Purpose}
@@ -18,12 +16,6 @@ import java.util.UUID
 import scala.concurrent.{ExecutionContext, Future}
 
 object ReadModelQueries {
-
-  def getEServices(offset: Int, limit: Int)(implicit
-    ec: ExecutionContext,
-    readModelService: ReadModelService
-  ): Future[Seq[CatalogItem]] =
-    readModelService.find[CatalogItem](ApplicationConfiguration.catalogCollection, Filters.empty(), offset, limit)
 
   def getTenant(
     tenantId: UUID
