+++ padigitale-report-generator/src/main/resources/application.conf
@@ -1,13 +1,9 @@
 report-generator {
   storage {
-    agreements {
-      container = ${AGREEMENTS_CONTAINER}
-      csv {
-        storage-path = ${AGREEMENTS_CSV_STORAGE_PATH}
-      }
-      json {
-        storage-path = ${AGREEMENTS_JSON_STORAGE_PATH}
-      }
+    interfaces-container = ${INTERFACES_CONTAINER}
+    pa-digitale {
+      container = ${PA_DIGITALE_CONTAINER}
+      storage-path = ${PA_DIGITALE_STORAGE_PATH}
     }
   }
   read-model {
@@ -16,9 +12,8 @@ report-generator {
        name = ${READ_MODEL_DB_NAME}
     }
     collections {
+        eservices = ${ESERVICES_COLLECTION_NAME}
         tenants = ${TENANTS_COLLECTION_NAME}
-        agreements = ${AGREEMENTS_COLLECTION_NAME}
-        purposes = ${PURPOSES_COLLECTION_NAME}
     }
   }
 }
