+++ padigitale-report-generator/src/main/resources/logback.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration debug="false">
+
+    <appender name="CONSOLE" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder class="ch.qos.logback.core.encoder.LayoutWrappingEncoder">
+            <layout class="it.pagopa.interop.commons.logging.LoggerLayout"/>
+        </encoder>
+    </appender>
+
+    <appender name="ASYNC" class="ch.qos.logback.classic.AsyncAppender">
+        <queueSize>8192</queueSize>
+        <neverBlock>true</neverBlock>
+        <appender-ref ref="CONSOLE"/>
+    </appender>
+
+    <logger name="it.pagopa.interop.padigitalereportgenerator" level="INFO"/>
+    <logger name="org.mongodb" level="WARN"/>
+
+    <root name="application" level="ERROR">
+        <appender-ref ref="ASYNC"/>
+    </root>
+
+</configuration>
