+++ padigitale-report-generator/src/main/scala/it/pagopa/interop/padigitalereportgenerator/report/Metric.scala
@@ -1,4 +1,4 @@
-package it.pagopa.interop.metricsreportgenerator.report
+package it.pagopa.interop.padigitalereportgenerator.report
 
 import io.circe._
 import io.circe.generic.semiauto._
