+++ padigitale-report-generator/src/main/scala/it/pagopa/interop/padigitalereportgenerator/util/Error.scala
@@ -1,0 +1,9 @@
+package it.pagopa.interop.padigitalereportgenerator.util
+
+import java.util.UUID
+
+object Error {
+  final case class TenantNotFound(tenantId: UUID) extends Exception(s"Tenant ${tenantId.toString} not found")
+  final case class MissingActivationTimestamp(descriptorId: UUID)
+      extends Exception(s"Descriptor ${descriptorId.toString} without activatedAt field")
+}
