+++ project/Dependencies.scala
@@ -144,13 +144,19 @@ object Dependencies {
       pagopa.parser        % Compile
     )
 
-    lazy val utils: Seq[ModuleID] = Seq(
-      // For making Java 12 happy
-      "javax.annotation" % "javax.annotation-api" % "1.3.2" % "compile",
-      //
-      logback.classic    % Compile,
-      pagopa.commons     % Compile,
-      pagopa.file        % Compile
+    val paDigitaleReportGenerator: Seq[ModuleID] = Seq(
+      cats.core            % Compile,
+      circe.core           % Compile,
+      circe.generic        % Compile,
+      logback.classic      % Compile,
+      mongodb.scalaDriver  % Compile,
+      pagopa.catalogModels % Compile,
+      pagopa.tenantModels  % Compile,
+      pagopa.commons       % Compile,
+      pagopa.cqrs          % Compile,
+      pagopa.file          % Compile,
+      pagopa.parser        % Compile
     )
+
   }
 }
